Advanced TicketBot/requirements.txt

discord.py>=2.3.2
aiosqlite>=0.19.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
